+++ runtime/cere/src/lib.rs
@@ -134,10 +134,10 @@ pub const VERSION: RuntimeVersion = RuntimeVersion {
 	// and set impl_version to 0. If only runtime
 	// implementation changes and behavior does not, then leave spec_version as
 	// is and increment impl_version.
-	spec_version: 53001,
+	spec_version: 53002,
 	impl_version: 0,
 	apis: RUNTIME_API_VERSIONS,
-	transaction_version: 16,
+	transaction_version: 17,
 	state_version: 0,
 };
 
@@ -1135,6 +1135,9 @@ impl pallet_ddc_clusters::Config for Runtime {
 	type StakerCreator = pallet_ddc_staking::Pallet<Runtime>;
 	type Currency = Balances;
 	type WeightInfo = pallet_ddc_clusters::weights::SubstrateWeight<Runtime>;
+	type MinErasureCodingRequiredLimit = ConstU32<4>;
+	type MinErasureCodingTotalLimit = ConstU32<6>;
+	type MinReplicationTotalLimit = ConstU32<3>;
 }
 
 impl pallet_ddc_nodes::Config for Runtime {
@@ -1323,6 +1326,7 @@ pub mod migrations {
 
 	/// Unreleased migrations. Add new ones here:
 	pub type Unreleased = (
+        pallet_ddc_clusters::migration::MigrateToV1<Runtime>,
 		pallet_contracts::migration::Migration<Runtime>,
 		pallet_referenda::migration::v1::MigrateV0ToV1<Runtime>,
 		// Gov v1 storage migrations